Rename v1.router to router/index
